refactor: 자동차의 위치를 받는 변수를 String형과 int형 두 개로 리팩토링
- Car 클래스
	- position -> positionString
	- positionInt 추가
- Game 클래스
	- getPosition -> getPositionString
- Winner 클래스
	- getPosition.length() -> getPositionInt
