+++ src/main/java/racingcar/domain/Winner.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.domain;
 
 import racingcar.domain.Car;
 
@@ -8,20 +8,24 @@ import java.util.List;
 public class Winner {
     public static int findFarthestCarPosition(List<Car> carList) {
         int farthestCarPosition = 0;
+
         for (Car car : carList) {
             // car.get().len() 수정 ->  Car 클래스에 position int로 받는거랑 str으로 저장 둘 다?
-            farthestCarPosition = Math.max(farthestCarPosition, car.getPosition().length());
+            farthestCarPosition = Math.max(farthestCarPosition, car.getPositionInt());
         }
+
         return farthestCarPosition;
     }
 
     public static List<String> winnerList(List<Car> carList, int farthestCarPosition) {
         List<String> winner = new ArrayList<>();
+
         for (Car car : carList) {
-            if (car.getPosition().length() == farthestCarPosition) {
+            if (car.getPositionInt() == farthestCarPosition) {
                 winner.add(car.getName());
             }
         }
+
         return winner;
     }
 }
